Expand mypy to whole project and fix all type errors
CI/local mypy was limited to quant_rl (with a large exclude list), so
mt5_trading, scripts, tests, backtest and the root demo/live scripts
were never type-checked. pyproject.toml now lists every package plus the
root scripts under [tool.mypy], and the exclude regex is gone.

Fixed the ~477 errors this surfaced:
- quant_rl: lazy cfg/Any typing in make_env, typed callbacks list,
  CheckpointCallback.save_path=str(model_dir), wandb added to
  ignore_missing_imports, scipy stubs, pandas/Figure casts in plots
- scripts: report_g3/discover_symbols/train_vae/run_baseline_eval fully
  annotated (dict[str, Any], Dataset[T], DataLoader[T], cast helpers)
- mt5_trading: strategy ABC uses TYPE_CHECKING for Signal (breaks the
  runtime chain into MetaTrader5, which is Windows-only); robots and
  manager got return types and dict[str, Any] annotations
- tests: dict/ndarray/tuple type args, get_indexer(pd.Index([...])),
  cast(pd.Timestamp, ...) for Hashable indices, protocol-typed stubs
- root scripts: demo_trading/live_trading null-safe credential guards

CI: ci-scripts/mypy and docker-compose typecheck service now run plain
mypy (project-wide via pyproject.toml) instead of mypy quant_rl.

Final: mypy 0 errors in 163 files, ruff clean, 351 tests pass.

+++ tests/test_models/test_encoders.py
@@ -2,6 +2,7 @@
 
 from __future__ import annotations
 
+import numpy as np
 import torch
 from gymnasium import spaces
 
@@ -15,8 +16,8 @@ class TestTCNEncoder:
         """Test TCNEncoder with proper observation space."""
         observation_space = spaces.Dict(
             {
-                "seq": spaces.Box(low=-1.0, high=1.0, shape=(60, 64), dtype=float),
-                "account": spaces.Box(low=-1.0, high=1.0, shape=(5,), dtype=float),
+                "seq": spaces.Box(low=-1.0, high=1.0, shape=(60, 64), dtype=np.float32),
+                "account": spaces.Box(low=-1.0, high=1.0, shape=(5,), dtype=np.float32),
             }
         )
 
@@ -33,8 +34,8 @@ class TestTCNEncoder:
         """Test TCNEncoder forward pass."""
         observation_space = spaces.Dict(
             {
-                "seq": spaces.Box(low=-1.0, high=1.0, shape=(60, 64), dtype=float),
-                "account": spaces.Box(low=-1.0, high=1.0, shape=(5,), dtype=float),
+                "seq": spaces.Box(low=-1.0, high=1.0, shape=(60, 64), dtype=np.float32),
+                "account": spaces.Box(low=-1.0, high=1.0, shape=(5,), dtype=np.float32),
             }
         )
 
@@ -64,8 +65,8 @@ class TestTransformerEncoder:
         """Test TransformerEncoder with proper observation space."""
         observation_space = spaces.Dict(
             {
-                "seq": spaces.Box(low=-1.0, high=1.0, shape=(60, 64), dtype=float),
-                "account": spaces.Box(low=-1.0, high=1.0, shape=(5,), dtype=float),
+                "seq": spaces.Box(low=-1.0, high=1.0, shape=(60, 64), dtype=np.float32),
+                "account": spaces.Box(low=-1.0, high=1.0, shape=(5,), dtype=np.float32),
             }
         )
 
@@ -82,8 +83,8 @@ class TestTransformerEncoder:
         """Test TransformerEncoder forward pass."""
         observation_space = spaces.Dict(
             {
-                "seq": spaces.Box(low=-1.0, high=1.0, shape=(60, 64), dtype=float),
-                "account": spaces.Box(low=-1.0, high=1.0, shape=(5,), dtype=float),
+                "seq": spaces.Box(low=-1.0, high=1.0, shape=(60, 64), dtype=np.float32),
+                "account": spaces.Box(low=-1.0, high=1.0, shape=(5,), dtype=np.float32),
             }
         )
 
@@ -111,8 +112,8 @@ class TestGRUEncoder:
         """Test GRUEncoder with proper observation space."""
         observation_space = spaces.Dict(
             {
-                "seq": spaces.Box(low=-1.0, high=1.0, shape=(128, 64), dtype=float),
-                "account": spaces.Box(low=-1.0, high=1.0, shape=(5,), dtype=float),
+                "seq": spaces.Box(low=-1.0, high=1.0, shape=(128, 64), dtype=np.float32),
+                "account": spaces.Box(low=-1.0, high=1.0, shape=(5,), dtype=np.float32),
             }
         )
 
@@ -129,8 +130,8 @@ class TestGRUEncoder:
         """Test GRUEncoder forward pass."""
         observation_space = spaces.Dict(
             {
-                "seq": spaces.Box(low=-1.0, high=1.0, shape=(128, 64), dtype=float),
-                "account": spaces.Box(low=-1.0, high=1.0, shape=(5,), dtype=float),
+                "seq": spaces.Box(low=-1.0, high=1.0, shape=(128, 64), dtype=np.float32),
+                "account": spaces.Box(low=-1.0, high=1.0, shape=(5,), dtype=np.float32),
             }
         )
 
@@ -154,8 +155,8 @@ class TestGRUEncoder:
         """Test GRUEncoder has trainable parameters."""
         observation_space = spaces.Dict(
             {
-                "seq": spaces.Box(low=-1.0, high=1.0, shape=(128, 64), dtype=float),
-                "account": spaces.Box(low=-1.0, high=1.0, shape=(5,), dtype=float),
+                "seq": spaces.Box(low=-1.0, high=1.0, shape=(128, 64), dtype=np.float32),
+                "account": spaces.Box(low=-1.0, high=1.0, shape=(5,), dtype=np.float32),
             }
         )
 
@@ -168,8 +169,8 @@ class TestGRUEncoder:
         """Test GRUEncoder with default parameters."""
         observation_space = spaces.Dict(
             {
-                "seq": spaces.Box(low=-1.0, high=1.0, shape=(128, 64), dtype=float),
-                "account": spaces.Box(low=-1.0, high=1.0, shape=(5,), dtype=float),
+                "seq": spaces.Box(low=-1.0, high=1.0, shape=(128, 64), dtype=np.float32),
+                "account": spaces.Box(low=-1.0, high=1.0, shape=(5,), dtype=np.float32),
             }
         )
 
@@ -190,8 +191,8 @@ class TestEncoderComparison:
         """Test that all encoders produce the same output shape."""
         observation_space = spaces.Dict(
             {
-                "seq": spaces.Box(low=-1.0, high=1.0, shape=(60, 64), dtype=float),
-                "account": spaces.Box(low=-1.0, high=1.0, shape=(5,), dtype=float),
+                "seq": spaces.Box(low=-1.0, high=1.0, shape=(60, 64), dtype=np.float32),
+                "account": spaces.Box(low=-1.0, high=1.0, shape=(5,), dtype=np.float32),
             }
         )
 
@@ -222,8 +223,8 @@ class TestEncoderComparison:
         """Test that all encoders accept dict input."""
         observation_space = spaces.Dict(
             {
-                "seq": spaces.Box(low=-1.0, high=1.0, shape=(60, 64), dtype=float),
-                "account": spaces.Box(low=-1.0, high=1.0, shape=(5,), dtype=float),
+                "seq": spaces.Box(low=-1.0, high=1.0, shape=(60, 64), dtype=np.float32),
+                "account": spaces.Box(low=-1.0, high=1.0, shape=(5,), dtype=np.float32),
             }
         )
 
@@ -245,8 +246,8 @@ class TestEncoderComparison:
         """Encoder output must depend on the sequence input."""
         observation_space = spaces.Dict(
             {
-                "seq": spaces.Box(low=-1.0, high=1.0, shape=(60, 64), dtype=float),
-                "account": spaces.Box(low=-1.0, high=1.0, shape=(5,), dtype=float),
+                "seq": spaces.Box(low=-1.0, high=1.0, shape=(60, 64), dtype=np.float32),
+                "account": spaces.Box(low=-1.0, high=1.0, shape=(5,), dtype=np.float32),
             }
         )
         encoder = TCNEncoder(
@@ -268,8 +269,8 @@ class TestEncoderComparison:
         """Encoder output must depend on the account input."""
         observation_space = spaces.Dict(
             {
-                "seq": spaces.Box(low=-1.0, high=1.0, shape=(60, 64), dtype=float),
-                "account": spaces.Box(low=-1.0, high=1.0, shape=(5,), dtype=float),
+                "seq": spaces.Box(low=-1.0, high=1.0, shape=(60, 64), dtype=np.float32),
+                "account": spaces.Box(low=-1.0, high=1.0, shape=(5,), dtype=np.float32),
             }
         )
         encoder = TransformerEncoder(
@@ -291,8 +292,8 @@ class TestEncoderComparison:
         """Same input must produce identical output."""
         observation_space = spaces.Dict(
             {
-                "seq": spaces.Box(low=-1.0, high=1.0, shape=(60, 64), dtype=float),
-                "account": spaces.Box(low=-1.0, high=1.0, shape=(5,), dtype=float),
+                "seq": spaces.Box(low=-1.0, high=1.0, shape=(60, 64), dtype=np.float32),
+                "account": spaces.Box(low=-1.0, high=1.0, shape=(5,), dtype=np.float32),
             }
         )
         encoder = TCNEncoder(
@@ -311,8 +312,8 @@ class TestEncoderComparison:
         """Same input must produce identical output."""
         observation_space = spaces.Dict(
             {
-                "seq": spaces.Box(low=-1.0, high=1.0, shape=(60, 64), dtype=float),
-                "account": spaces.Box(low=-1.0, high=1.0, shape=(5,), dtype=float),
+                "seq": spaces.Box(low=-1.0, high=1.0, shape=(60, 64), dtype=np.float32),
+                "account": spaces.Box(low=-1.0, high=1.0, shape=(5,), dtype=np.float32),
             }
         )
         encoder = TransformerEncoder(
